Use dict and set instead of HashSet
GraphNode was storing a HashSet of GraphEdge, now stores a dict indexing each
GraphEdge by its value. Renamed attribute from edges to edgemap.

Graph was storing a HashSet of GraphNode, now stores a dict indexing each
GraphNode by its value. Renamed attribute from hashmap to nodemap.

Graph.nearest_unexplored_edge was storing touched nodes as a HashSet, when a
normal set worked just fine.

Deleted hashset.py as it is no longer and will never be used. It was a
misguided implementation anyway (hashes should never be used as keys in a
dict, and the prototype access model is ill-advised)