Stop storing GraphEdge.explored
GraphEdge used to use the 'explored' attribute to keep track of whether the
'target' attribute had been set. Now, it just sets target to None by default,
and checks if it is not equal to None. Removed the set_target method as it was
no longer necessary, and added the is_explored method