Removed GraphEdge
The function of GraphEdge is to store the value and target of an edge. This
does not actually require a whole object, and can be implemented as a dict in
GraphNode. GraphNode.edgemap is now a mapping from edge values to targets,
instead of from edge values to GraphEdge objects.
GraphNode.get_explored_edges() and get_unexplored_edges() now iterate through
edge values instead of GraphEdge objects, as they should anyway.