Fix Graph.add and magic methods
Graph.add used to silently throw away the given item if it already existed in
the graph. Now it throws an exception in that case, because that check should
be done explicitly by graph's user when necessary. Crawler now reflects the
new usage.

Graph.__contains__ used to erronously compare the given item to a GraphNode
object by calling __contains__ on nodemap.values(). It now compares it to the
keys of nodemap, which are the items stored in all the GraphNodes, as expected.
Graph.__iter__ and Graph.__str__ were also updated to deal directly with items
instead of GraphNodes, but these were only changes in methodology, not
functionality.