Add Click features to crawler
Instead of using button objects as graph edges, the crawler now uses
click objects as graph edges. This means that the graph edges are
unique for buttons with different settings. Added functionality to merge
buttons and screens. Did NOT add functionality to save information about
effects on state variables in the click object. The crawler now
understands that clicking a button with a different setting can bring it
to a different screen, but it does not yet understand how those settings
are set.