Extend run_crawler.py to run per-function
run_crawler.py now runs it's crawl and help window steps on the setting screens of each function in a stored list. Many functions are currently commented out because the program is not yet able to handle them. The pickle saving and loading functionality was changed so that filenames now include the name of the function. A final step was added to run after each function is analyzed that uses the saved pickles to build a master csv file with all the saved help caption data.