Redesign merge button/screen functions
crawler.merge_buttons used to look at the entire set of stored settings
for both given buttons and merge the two. However, it is only ever used
where the source button has a single setting. For greater usefullness
in the future, merge_buttons was changed to only compare the current
setting of each button. This means the set object returned by
merge_buttons is limited to one item. This also affects the
functionality of merge_screens.